Output rgba. This removes the braindead accumulation into a 32bit value.
git-svn-id: svn://svn.netsurf-browser.org/trunk@9573 d004b94f-0d10-0410-95b5-d60e4a4a7a9a

+++ libnspng/src/internal.h
@@ -147,7 +147,7 @@ struct nspng_ctx {
 	uint8_t prev_pixel[8];
 	uint8_t *src_scanline;
 
-	uint32_t *rowbuf;
+	uint8_t *rowbuf;
 
 	z_stream zlib_stream;
